physical-ai-humanoid-robotics-textbook/RAG-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
qdrant-client==1.7.0
openai==1.3.7
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
tiktoken==0.5.2
sentence-transformers==2.2.2
torch==2.5.1
transformers==4.45.2